feat: index페이지 리다이렉션 추가
